+++ src/impl/model/ModelSix/SampleScheduler.h
@@ -27,8 +27,8 @@ namespace transmission_nets::impl::ModelSix {
     SampleScheduler<T, Engine, Scheduler>::SampleScheduler(std::shared_ptr<State> state, std::shared_ptr<T> target, std::shared_ptr<Engine> r, int samplesPerStep) : state_(std::move(state)), target_(std::move(target)), r_(r), scheduler_(r_, samplesPerStep) {
         using namespace core::samplers;
 
-        long double totalInfections = state_->infections.size();
-        long double totalLoci       = state_->loci.size();
+        double totalInfections = state_->infections.size();
+        double totalLoci       = state_->loci.size();
 
         scheduler_.registerSampler({.sampler = std::make_unique<ConstrainedContinuousRandomWalk<T, Engine>>(state_->geometricGenerationProb, target_, 0.0, 1.0, r, 1, .01, 2),
                                     .id      = "Geometric Generation Prob",
